+++ source/renderer/app/actions/wallet-backup-actions.js
@@ -5,7 +5,9 @@ import Action from './lib/Action';
 
 export default class WalletBackupActions {
   startWalletBackup: Action<any> = new Action();
-  initiateWalletBackup: Action<{ recoveryPhrase: string[] }> = new Action();
+  initiateWalletBackup: Action<{
+    recoveryPhrase: Array<string>,
+  }> = new Action();
   acceptPrivacyNoticeForWalletBackup: Action<any> = new Action();
   continueToRecoveryPhraseForWalletBackup: Action<any> = new Action();
   addWordToWalletBackupVerification: Action<{
@@ -18,4 +20,7 @@ export default class WalletBackupActions {
   restartWalletBackup: Action<any> = new Action();
   cancelWalletBackup: Action<any> = new Action();
   finishWalletBackup: Action<any> = new Action();
+  // Recovery phrase confirmation dialog actions
+  checkRecoveryPhrase: Action<{ recoveryPhrase: Array<string> }> = new Action();
+  resetRecoveryPhraseCheck: Action<any> = new Action();
 }