+++ storybook/stories/SidebarWalletsMenu.stories.js
@@ -4,6 +4,10 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import StoryDecorator from './support/StoryDecorator';
 import SidebarWalletsMenu from '../../source/renderer/app/components/sidebar/wallets/SidebarWalletsMenu';
+import {
+  WalletRecoveryPhraseVerificationStatuses,
+  WalletRecoveryPhraseVerificationTypes,
+} from '../../source/renderer/app/stores/WalletsStore';
 
 storiesOf('SidebarWalletsMenu', module)
   .addDecorator(story => (
@@ -36,6 +40,12 @@ storiesOf('SidebarWalletsMenu', module)
           isRestoreActive: false,
           restoreProgress: 0,
           isLegacy: false,
+          createdAt: new Date(),
+          recoveryPhraseVerificationDate: new Date(),
+          recoveryPhraseVerificationStatus:
+            WalletRecoveryPhraseVerificationStatuses.OK,
+          recoveryPhraseVerificationStatusType:
+            WalletRecoveryPhraseVerificationTypes.NEVER_CHECKED,
         },
         {
           id: '2',
@@ -45,6 +55,12 @@ storiesOf('SidebarWalletsMenu', module)
           isRestoreActive: false,
           restoreProgress: 0,
           isLegacy: false,
+          createdAt: new Date(),
+          recoveryPhraseVerificationDate: new Date(),
+          recoveryPhraseVerificationStatus:
+            WalletRecoveryPhraseVerificationStatuses.OK,
+          recoveryPhraseVerificationStatusType:
+            WalletRecoveryPhraseVerificationTypes.NEVER_CHECKED,
         },
         {
           id: '3',
@@ -54,6 +70,12 @@ storiesOf('SidebarWalletsMenu', module)
           isRestoreActive: false,
           restoreProgress: 0,
           isLegacy: false,
+          createdAt: new Date(),
+          recoveryPhraseVerificationDate: new Date(),
+          recoveryPhraseVerificationStatus:
+            WalletRecoveryPhraseVerificationStatuses.OK,
+          recoveryPhraseVerificationStatusType:
+            WalletRecoveryPhraseVerificationTypes.NEVER_CHECKED,
         },
         {
           id: '4',
@@ -63,6 +85,12 @@ storiesOf('SidebarWalletsMenu', module)
           isRestoreActive: false,
           restoreProgress: 0,
           isLegacy: false,
+          createdAt: new Date(),
+          recoveryPhraseVerificationDate: new Date(),
+          recoveryPhraseVerificationStatus:
+            WalletRecoveryPhraseVerificationStatuses.OK,
+          recoveryPhraseVerificationStatusType:
+            WalletRecoveryPhraseVerificationTypes.NEVER_CHECKED,
         },
       ]}
       isActiveWallet={id => id === '2'}