+++ storybook/stories/support/StoryLayout.js
@@ -22,6 +22,7 @@ import Sidebar from '../../../source/renderer/app/components/sidebar/Sidebar';
 import SidebarLayout from '../../../source/renderer/app/components/layout/SidebarLayout';
 import menuIconOpened from '../../../source/renderer/app/assets/images/menu-opened-ic.inline.svg';
 import menuIconClosed from '../../../source/renderer/app/assets/images/menu-ic.inline.svg';
+import { WalletRecoveryPhraseVerificationStatuses } from '../../../source/renderer/app/stores/WalletsStore';
 
 export type StoriesProps = {
   wallets: Array<Wallet>,
@@ -114,6 +115,8 @@ export default class StoryLayout extends Component<Props> {
       isRestoreActive: get(wallet, 'syncState.tag', 'synced') === 'restoring',
       restoreProgress: get(wallet, 'syncState.data.percentage.quantity', 0),
       isLegacy: wallet.isLegacy,
+      recoveryPhraseVerificationStatus:
+        WalletRecoveryPhraseVerificationStatuses.OK,
     }));
 
   getSidebarMenus = (