+++ storybook/stories/support/StoryProvider.js
@@ -22,6 +22,7 @@ export const WALLETS = [
     passwordUpdateDate: new Date(),
     syncState: { data: null, tag: 'synced' },
     isLegacy: false,
+    recoveryPhraseVerificationDate: new Date(),
   },
   {
     id: '1',
@@ -34,6 +35,7 @@ export const WALLETS = [
       .toDate(),
     syncState: { data: null, tag: 'synced' },
     isLegacy: false,
+    recoveryPhraseVerificationDate: new Date(),
   },
   {
     id: '2',
@@ -44,6 +46,9 @@ export const WALLETS = [
     passwordUpdateDate: new Date(),
     syncState: { data: null, tag: 'synced' },
     isLegacy: true,
+    recoveryPhraseVerificationDate: moment()
+      .subtract(200, 'days')
+      .toDate(),
   },
   {
     id: '3',
@@ -70,6 +75,9 @@ export const WALLETS = [
       tag: 'restoring',
     },
     isLegacy: false,
+    recoveryPhraseVerificationDate: moment()
+      .subtract(400, 'days')
+      .toDate(),
   },
 ];